+++ dmr/security/jwt/__init__.py
@@ -14,4 +14,6 @@ except ImportError:  # pragma: no cover
 from dmr.security.jwt.auth import JWTAsyncAuth as JWTAsyncAuth
 from dmr.security.jwt.auth import JWTSyncAuth as JWTSyncAuth
 from dmr.security.jwt.auth import request_jwt as request_jwt
+from dmr.security.jwt.cookie import CookieJWTAsyncAuth as CookieJWTAsyncAuth
+from dmr.security.jwt.cookie import CookieJWTSyncAuth as CookieJWTSyncAuth
 from dmr.security.jwt.token import JWToken as JWToken